requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.4
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==22.0.0
idna==3.6
importlib-metadata==6.8.0
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.1
MarkupSafe==2.1.5
mysqlclient==2.2.0
openapi-codec==1.3.2
packaging==23.2
phonenumbers==8.13.26
Pillow==10.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
swagger-ui-py==23.9.23
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.0
zipp==3.17.0
